🔵 factorize types
