+++ packages/backend/src/controller/types.ts
@@ -1,0 +1,8 @@
+interface ErrorJSON {
+  message: string
+}
+
+interface ValidateErrorJSON {
+  message: 'Validation failed'
+  details: { [name: string]: unknown }
+}
